Added ShaderAsset
This stores information about a single shader as well as it's compiled binary code.

+++ include/Rendering/Vulkan/VKPipeline.h
@@ -24,12 +24,12 @@ namespace puffin::rendering::util
 
 			// find what the size of the file is by looking up the location of the cursor
 			// because the cursor is at the end, it gives the size directly in bytes
-			const size_t fileSize = (size_t)file.tellg();
+			const size_t fileSize = file.tellg();
 
 			// spirv expects the buffer to be on uint32, so make sure to reserve an int vector big enough for the entire file
 			std::vector<uint32_t> buffer(fileSize / sizeof(uint32_t));
 
-			// put fiel cursor at beginning
+			// put file cursor at beginning
 			file.seekg(0);
 
 			// load entire file into buffer
